Beginning automatic git commiting
Adding flashbake, an automated script that will push changes on
a regular basis. Commit messages will just be the weather, and can
be filtered based on that.

This is the crontab line:
15 21	* * *	jay	flashbake /home/jay/Documents/PhD-Working 60 > /dev/null

and the tool is from here:
https://github.com/commandline/flashbake/wiki

I will still make manual commits or tags when an informative commit message is needed.