VWM Chapter intro to James
Finished rough draft of VWM and sent to James. Prisms intro is ready to be
worked on. Added git status to the git-wc.sh script, as it's becoming a
way to monitor progress with "watch". I might also add a grep for FIXME:,
TODO: and 'tk' instances to either count them or display them. Also deleted
the pull_readme.sh script as I've given up on having seperate private and public
repositories for the phd. I'll just use this one and switch it to public on github
when I publish.