Issue 2.2.3
The experiment doesn't directly test the role of rehearsal, but
replicates working memory deficits while ruling out, specifically,
an inability to utilize rehearsal using spatial attention,
as the cause of the deficit. I tried to clarify the introduction passage
regarding this, and added a paragraph to the general discussion
acknowledging the need for further research to flesh out this idea.