Issue 2.2.11
Normality of the precision data is borderline when subjected to
diagnostic plots. Rather than argue that it's acceptable, I re-ran the
analysis with log transformed data (which is much more normal). The
anova then comes out on the razors-edge of being non-significant, so I
add post-hoc comparisons. The overall message doesn't change, but it's
more thorough.